+++ crates/rust-client/src/store/note_record/input_note_record/states/consumed_authenticated_local.rs
@@ -1,6 +1,7 @@
 use alloc::string::ToString;
 
 use miden_protocol::Word;
+use miden_protocol::account::AccountId;
 use miden_protocol::block::{BlockHeader, BlockNumber};
 use miden_protocol::note::{NoteId, NoteInclusionProof, NoteMetadata};
 use miden_protocol::transaction::TransactionId;
@@ -39,6 +40,7 @@ impl NoteStateHandler for ConsumedAuthenticatedLocalNoteState {
     fn consumed_externally(
         &self,
         _nullifier_block_height: BlockNumber,
+        _consumer_account: Option<AccountId>,
     ) -> Result<Option<InputNoteState>, NoteRecordError> {
         Ok(None)
     }